Allow viewing closed meeting when meeting dossier is closed.
A user with edit permissions on the meeting would not be able
to view the meeting if the meeting dossier is closed. This is
to avoid problems with buttons in the meeting view that do not
work because of lack of edit permissions on the meeting dossier.
Nevertheless when the meeting is closed, it should be visible
for a user with edit permissions even when the meeting dossier
is closed.