Register ChoiceFieldDeserializer using overrides instead of configure ZCML:
This ensures that it will be loaded with the currently pinned version
of plone.restapi (which doesn't have it's own ChoiceFieldDeserializer
yet) and will also not cause conflicting configurations with
plone.restapi:master, which now does have a ChoiceFieldDeserializer.

That way we make sure that GEVER tests against plone.restapi are
green again and don't shadow problems that might be introduced
in the upcoming restapi sprint.