Resolve view: Validate preconditions early:
We do this so we don't redirect to the archive form
(if filing number feature enabled) in a case where
it will fail anyway because of violated preconditions.

This will validate preconditions *twice* though (the second
time via resolve_manager.resolve()). This should eventually be
cleaned up so we don't unnecessarily validate preconditions
multiple times.