Schema dumps: Add workaround for getTaskTypeVocabulary() not
being recognized as order-independent:

The getTaskTypeVocabulary() factory destroys the reference
to the underlying VDEX vocabulary by incorrectly wrapping it.
Therefore the detection for order significance doesn't work.

That means that in the schema dumps, the terms for this vocabulary
don't get sorted alphabetically (even though they should, because
the vocab *doesn't* have an inherent sort order, and therefore returns
terms in variable order). This leads to flaky tests if we don't
force order_significant = False for this vocab.