Use separte NewResponsible schema for forwarding reasign transition.
The ForwardingReassignTransitionExtender had used the same schema as
the ReassignTransitionExtender for tasks. That extender has the
INewResponsibleSchema associated with it, which configures the
AllUsersInboxesAndTeamsSourceBinder (used for the new responsible)
with only_current_inbox=True. That is correct for tasks, but not
forwardings. When reassigning tasks, the org unit is not allowed to
change, but forwardings can be refused and assigned to the inbox of a
different org unit.