Avoid reindexing 'created' during IObjectCopiedEvent:
If an object has been copy & pasted, we need to reindex some fields.
The IObjectCopiedEvent is too early to do that though, because at
that point the object doesn't have a full AQ chain yet. We therefore
just mark it during IObjectCopiedEvent, and do the reindexing later
during IObjectAddedEvent.

This fixes copy & pasting in Solr-enabled deployments.