Dont require plone site in exception formatter.
If the plone site does not exist, e.g. for errors that happen while we install
the plone site in the first place, formatting the exception would fail with
a further exception, `CannotGetPortalError`.

We change the permission check to use the zope application and retrieve the
application from the request. Gracefully hide the trace in case of a missing
request or if the app cannot be retrieved.