Remove confusing --skip-solr option
The option only disables Solr in the reindex section.
But there are other places
(e.g. opengever.base.security.reindex_object_security_without_children)
where Solr is still called explicitly.
To avoid Solr indexing, it should be disabled in the registry.

Btw, the option is misnamed. If the option is provided, `skip_solr` has a
value of `False`.?! If the option is omitted, the value is `True`.