Enable c.indexing during tests, but patch it to not defer operations:
We need to enable collective.indexing during tests in order to
create Solr tests (because ftw.solr integrates itself by providing
an IIndexQueueProcessor).

So we need c.indexing's pluggable IndexQueueProcessor mechanism, but
we don't want the unpredictable behavior of reindexes that are deferred
until the end of the transaction during tests (or the queue getting
flushed at arbitrary points).

We therefore patch the IndexQueue's operation methods so that they
always immediately process the queue after being called.