Fix workspace participation restapi.
The new plone.restapi sends a dict having a token and a title for vocabularie values like the users. The frontend is using this new format as well. We have to update the participation implementation to use it as well to fix an issue where it was no longer possible to add participants.