Move language field to proposal content type.
The language field in the database is kept. We will sync data to it also in the
future. The primary data will now be stored on the plone content type though.

Also handle submitted proposal as not doing that would be more complicated.

The upgrade step can be made deferrable as setting the field immediately is not
strictly necessary.