Clearly separate proposal fields.
Separate fields so that it is clear which ones are used for both, submitted
proposal and proposal and which ones are only used on one side of the two
proposals. From now on it will be possible to add new fields to the proposal
side only. Currently this would not have been possible as the submitted proposal
schema interface inherits from the proposal schema.

We end up removing two fields from the submitted side for now. We will add more
fields in the near future that will go to the proposal side only, thus this acts
as a preparation for the next step.

We also remove the fields from all content objects, they will currently contain
persisted default values which are never used.