Separate event handlers for submitted proposal.
Separate event handlers for submitted proposal from event handlers for proposal.
We only have to handle modified events as the model instance is only added on
the proposal side and always must exist when a submitted proposal is created.
Also the set of attributes that is synced from the submitted proposal side is
distinct from the proposal attributes, thus we have a different event handler.