Shift permission checks around.
We no longer can rely on the modify permission as the plone workflow actively
removes modify for some states. We would like to be able to exit those states.
We move some modify checks to the submitted proposal side. Checking for the
modify permission will be moved to the workflow or the
ProposalTransitionController.