Receive scheduled transition via separate view.
Communicate to a "remote" proposal in the dossier that its corresponding
submitted proposal has been scheduled in a meeting. When a submitted proposal
has been scheduled we now want to trigger a workflow transition on the proposal
Do this via a separate view so that we can perform the workflow transition with
all other required actions (history entry, activity) at once.

We also drop the requirement that proposal history entries must be synced
completely and that uuid on local and remote side should be the same.