Introduce actual classes for reminder types:
This refactoring introduces a Reminder base class, and subclasses that define
specific reminder types. Instances of these classes are used to represent
specific reminders that may have reminder-instance specific dynamic parameters.

Static characteristics of the reminder types are
expressed via class level attributes, and the functions for calculating the
trigger date for the reminder are proper methods, which have access to the
reminder's parameters.

These classes are tracked in a registry of types, and the base class has a
factory method that is capable of constructing an instance of the appropriate
subclass, given an `option_type` string (by looking up the subclass in that
registry).