Handle (query)sources for multivalued zope.schema fields:
Multivalued fields don't directly have sources themselves, but
their items do.

We therefore need to handle that case in the @sources and
@querysources endpoints, and look for a source on the Choice
field used as the value_type for the multivalued field.

Similarly we need to construct appropriate URLs during
schema serialization for these types of fields: The Choice
fields used as the value_type for multivalued fields don't
usually have a name. We therefore omit their empty string
name from the URL, and instead construct URLs that look
like the source was actually on the multivalued fields
themselves.