Fix authorization handling when fetching the template_group_id.
With the latest change, we adjust the handling to extract the
templatelibrary id. But with this change, we mistakenly only adjusted the happy case,
but not the case when the first requests is answered with an 401,
because of a no longer valid token.

Fixes https://sentry.4teamwork.ch/sentry/onegov-gever/issues/52318/