Fix automatic start of additionally added sequential tasks.
The current implementation has not added the
IFromSequentialTasktemplate to additionally added sequential
tasks, which lead to an incorrect tasktemplate_predecessor
indexing. Which meant they weren't started automatically, when closing
the previous task.

No need for an upgradestep, the problem in the production was
solved by manually opening or skipping the task.