Tests: Fix fixture caching for Solr tests:
The ContentFixtureWithSolrLayer also sets up a GEVER fixture, but also needs
to dump and load the Solr database. On the other hand, the regular
ContentFixtureLayer MUST NOT attempt to dump / load Solr.

We therefore need to isolate their caches, so they can be dumped and loaded
separately. We do this by introducing a `cache_label` for the DBCacheManager,
which will store its stacks in a subdirectory named after the cache label
to keep them apart.