Adapt invitation storage, add email and status.
We want to make invitations email based only. We now require an email as
non optional parameter. An email based user-lookup is triggered upon
invite. The user is optional though and could be created in a later
step. Listings are now email based, too.

Also we add a workflow state to be able to store reject/accepts of
invitations. It is currently only allowed to have one pending
invitation.