Await solr to be restored when tearing down test layers
Instead of waiting for solr to have restored a backup when setting up the
testlayer, properly tear it down by waiting for solr to be finished restoring.

A symptom of this issue was, that solr logged:
"Restore in progress. Cannot run multiple restore operationsfor the same core"

See https://ci.4teamwork.ch/builds/302050/tasks/496604