Testserver: introduce BUILDOUT_DIR environment variable.
When the testserver is installed with a buildout located outside of opengever.core,
the bin-scripts (e.g. solr) are not located within the opengever.core checkout or egg
directory.

With the new BUILDOUT_DIR environment variable a caller (such as e2e UI tests) can
configure the correct buildout dir so that the testserver has access to the scripts.