Uses newline character to break text for helptext in forms
As discussed in https://github.com/4teamwork/gever-ui/issues/976 the gever-ui is not able to format form hints using HTML. So change the </br> to a \n. To make the newline character visible in the opengever.core use `white-space: pre-wrap;`. See the `pre-wrap` specification in https://tympanus.net/codrops/css_reference/white-space/.