Use collaborative checkouts for WOPI view:
This ensures that collaborators are tracked, and other
users than the one that originally checked out the
document can check it in via the WOPI view.

This also lays the foundation to clearly recognize
documents checked out collaboratively, in order to
later restrict certain actions that can be done with
documents in that state.