Backdate AddHiddenFlagToAdminAndOrgUnit upgrade step.
The current admin unit is fetched every time the metadata of
an object is updated, because of the reference number indexer.
This will lead to failing upgrade steps if we do not update
the admin_units table first. So we backdate this upgrade step
to just before the oldest upgrade that has not been executed
on all client installations yet.