OfficeOnline: Only explicitly track *additional* collaborators
via the 'office_online_edit' view. These should only be added to
the collaborator list of the document is already a collaborative
checkout. Otherwise we accidentally turn an exclusive checkout
into a collaborative one, by adding a user to the collaborators
list.

(that list is used to determine whether a checkout is collaborative
or not).