WOPIView: Prevent exclusively checked out doc from being edited with Office Online:
If a document is checked out exclusively (instead of collaboratively,
or not at all), the WOPI view must now allow an attempt to edit this
document in Office Online.

These cases should be prevented in the UI whenever possible. But
this acts as a last line of defense to prevent data loss / conflicting
changes.

If this case ever occurs, Office Online will display an error message
saying that this document unfortunately can't be opened, and suggests
to open it in a read only mode. The read only mode doesn't seem to
currently work, but it's answered with another message saying the
server is overloaded, and then only offers to quit the Office Online
editing window.

So, not terribly graceful, but data loss is averted.