Extend the `meeting.json` with a `agenda_item_list` property
Exporting a meeting to a zip file will generate a `meeting.json`-file which contains properties with links to the generated documents and agenda items. The agenda_item_list has been included in the zip-file but has not been linnked in the `meeting.json`. This commit extends the `meeting.json` with the `agenda_item_list` document.