Fix `task_type_helper` to respect the current language for the ram-cache.
The task_type_helper is a helper function for listings to return the task type title based on a task_type. The used vocabulary is a vdex vocabulary and provides language specific task type titles. The function is decorated with a ram-cache but only with the task_type as a cache-key. That means, if the user changes the language, the cache key is still valid and the returned value will be in the old language. To fix this problem, we have to make the cache key langauge dependent.