Enhance the breadcrumbs API endpoint.
The endpoint now returns a summary of the breadcrumbs instead of id and title, including the review state and the portal type.

A new key `is_leafnode` indicates if a repository folder is a leaf node or not (i.e. has children of the type repository folder).

A new key `is_subdossier` indicates if a business case dossier is a subdossier (i.e. its parent is a business case dossier too).