Improve has_sametype_children indexer performance.
Skip all objects that are obviously documents based on their key. This
will prevent loading all document objects of a dossier into memory.

Also don't retrieve the keys in sort order as it is irrelevant for
indexing.

Highest performance impact for dossiers without any subdossiers but
with many documents where no documents will be loaded any more instead
of all of them.

This strategy could also be applied to other content types, but i have
chosen to stick to documents for now as they are the most common one.