Show proposal templates corresponding to the committee.
In commit 31b96c4ea3b395224d1307b5c11dccfb3898898e the field 'committee' was renamed to 'committee_oguid'. This had the consequence that a select field was renamed waht caused that form.widgets.proposal_template wasn't updated anymore when changing the value in the select box.

Also the vocabulary filterd not correctly also due the switch to ougid.

This behavior is fixed in this commit.