Drop sorting by sortable_author for solr listings.
We now use solr for all listings. Sorting by sortable_author does not
make sense for solr as there is no sortable_author in solr. In
`portal_catalog` we used a `ZCTextIndex` index for `document_author`
thus we also had to introduce the `sortable_author` index for proper
sorting. It could probably be dropped but that is out of scope for this
fix and will come separately.