Remove SearchableText indexer
There is a skins-file for the livesarch which will be taken if we do not call the view explicitly. This file does not provide solr and sets the contenttype to xml instead html. This will break the test as soon as we remove the SearchableText indexer from the plone catalog.