Avoid sending mails with From-Addresses other than our own:
We must not send mails on behalf of users by using their mail
address in the From: header, and therefore effectively spoofing
the sender address. Such mails will get rejected in modern mail
setups, e.g. when DKIM is in use.

Instead, we include the user's full name in the From: header,
but use the system-wide 'noreply' address of the GEVER deployment.