Make From: address behavior configurable for mails
sent on behalf of activity actors. The new default is to
use the system noreply address (to avoid sender address
spoofing), but by setting the 'send_with_actor_from_address'
flag to True, the legacy behavior can be restored where the
actor's email address is used directly in the From: header.