Add minimalistic log handler for sentry.
Explicitly log to sentry for two `ftw.solr` modules we want to monitor
well at the moment.

Not intended to stay around for long but rather a quick way to get
logging to sentry for potentially critical solr errors. Should go away
again once we have proper logging for non-exceptions to sentry.