Allow customizing group dn for invitations
By default invited users are added to the org unit's users group.
This group includes all users that have access to the site.
However often invited uses need be put into a dedicated group.
This allows configuring a dedicated group in the registry.
If not configured, it falls back to the org unit's users group.