Fix deadline of task templates no longer shown in tabular listing.
Since we switched the tabular listing view to Solr, the deadline of task templates was no longer shown. There is already a Solr field called `deadline`, which is a `pdate`. But the deadline of the task templates is an integer. So we store the value in a Solr field called `period` and render it in the tabular listing.