Disable the searchbox for document versions.
Until now, a searchbox was rendered on the document versions tabbed view. As far as we could find out, the searchbox never had any effect on the results shown in table.

The listing of the document versions is lazy and we do not see an easy way to implement the filtering of the versions by a searchable text and keeping this lazyness at the same time.

According to the product owner, we may simply disable the searchbox on the version tab.