Make "populate_filename_column_in_favorites" UpgradeStep more robust:
This upgrade step attempts to poplate the filename SQL column for favorites,
which is limited to 105 characters. If an object (specifically mails) is encountered
that for some reason still has a longer filename indexed in the catalog,
this upgrade step will fail.

This change makes it more robust by truncating any filename
to the max length before attempting to write it to SQL.