Also expire any __ac cookie that might have been issued by the
CookieAuthHelper on the Zope root (e.g. for zopemaster).

In a vanilla Plone, these get reset by the Plone session plugin's
reset credentials (as a side effect), since both their cookie names
are set to '__ac'.

But because we change the session plugin's cookie name to
'__ac_<plone_site_id>' when installing ftw.casauth, these names
diverge and the CookieAuthHelper's cookie doesn't get purged as
a side effect any more, so we need to do it ourselves.