Add support for several metadata_fields for real content listing objects.
The summary serializer should support the same metadata, whether
the object being summarised is a brain or a real object, so that
items and related_items get serialized in the same way.
OpengeverRealContentListingObject should therefore support the same
metadata_fields as the OpengeverCatalogContentListingObject. As a
first step here, we add support for some essential fields.