Enhance "@config" with "private_folder_path".
Expose the path to the user's private folder in the @config API endpoint. This serves as a feature flag for the GEVER UI too.

The reason for this addition is the fact, that the GEVER UI does not get the path to the private folder right in every case. Especially when the user id is an e-mail address, the id of the private folder is normalized (see the tests for an example).