Robustify adding favorites with invalid workflows.
When for some reason workflow state cannot be retrieved we still want to
be able to create the favorite. Thus make the code a bit more robust
for this non-critical information.

Also fix an upgrade that would attempt to write Missing.Value from the
catalog into an sql column which won't work due to missing serializers.