Fix state_var name for contact workflow.
The state variable was configured as "state". While this is perfectly
valid configuration there are various other parts of the software
stack that always expect the variable to be configured as
"review_state".

We fix the configuration and also provide a migration for the workflow
history (and current workflow state).

Also add a basic test for the api as it also relied on review_state
and previous to this fix returned `None` as state.