Nullify docprops instead of deleting them.
This will retain text type doc-properties which are None in the
word document while updating it.

If we remove the doc-property from the word document and the
doc-property is actually used in the document body manually updating
doc-properties in word will cause an obtrusive warning message being
rendered into the word document body in place of the doc-property.

With this change we can avoid that the warning message appears, instead
word will just render empty text, or "nothing".