Fix @workflow when no permissions in target state.
Fix executing workflow transitions via @workflow service when the
executing user has no permission in target state on the context for
which the transition is executed.

This is done by accessing `review_history` with elevated privileges, if
necessary.

There is an issue in the `WorkflowTransition` service when the user
performs a transition into a state where his access is revoked. The
service will raise a `WorkflowException` when attempting to access
`review_history` while rendering the response. This will prevent the
transition from being executed.

We work around this issue when we read the history with elevated privileges.