Fix docs only partially indexed in solr.
We think affected documents may have been not indexed in solr at all due
to a bug with date serialization in `ftw.solr`. The bug was only fixed
in https://github.com/4teamwork/ftw.solr/pull/158.

Subsequent edits of the document by users via plone or reindex
operations during upgrades would have partially indexed or restored the
document in solr, but only some of the fields as we use atomic updates.

We've seen that all affected documents do not have the created date set
as we never manually reindex it, as it never changes. Thus apply a
heuristic to detect broken solr documents. We assume that broken
documents do not have the field 'created' set to a value in solr.