Reindex modified containers after bundle import.
In the oggbundle import pipeline, object reindexing is deactivated
to avoid objects being reindexed several times during the import.
Every object is indexed just once in the pipeline, but containers
into wich objects are added were not reindex afterwards, leading
to a discrepency in the modified date on the object and in the
catalog. This in turn leads to problems with solr diff / sync.