Prevent deadlock when reassigning inter-admin-unit tasks:
This solves an edge case which can lead to a deadlock when
reassigning inter-admin-unit tasks. If the new responsible
doesn't have an entry in the 'watchers' SQL table yet, both
the predecessor and successor sides would try to add it for
the first time, and therefore deadlock on the to-be-inserted
row for the new watcher.

This is prevented by creating the new watcher (if required) at
the very beginning of this process, and immediately committing
the transaction, and starting a new one.