Fix ogds listing endpoints for undefined sort orders.
When sorting according to a non-unique column, response of
ogds listings could contain duplicate items and other items
could be missing. This is because the sort order of successive
queries is not guaranteed to be identical (except when specifically
sorting on a key with a unique value for each item).