Temporary disable lock check in tus-replace
The current Office Connector release (1.10.0) does not yet send the lock
token when replacing a file. The check for the correct lock token would
break uploading files with the current OC release.

We need a some time to ensure a wide rollout of an OC release with lock
token support. Thus the test verifying that a file can be replaced only
if the correct lock token was supplied, is skipped for the one year.