Manage groups with source groups plugin and not with ldap plugin.
As we now overwrite the @groups endpoint allowing also admin to manage
groups, we want to avoid mistakingly writing in the ldap. Managed groups
should all be source groups.
We now also need to find the created source groups, hence we need to
activate enumeration for the source_groups plugin.