Check group management and enumeration plugins in ogds-sync.
We make sure here that the site is setup correctly for the @groups
endpoint to work (i.e. groups are created as source_groups and
source groups can be enumerated).
While this is not necessary for the ogds-sync, it allows us
to make sure that the configuration is checked on a regular basis.