Migrate IDossierTemplate fields storage.
As IDossierTemplate inherited from IDossier, the fields were
stored in the annotations under the IDossier interface. As
we have now moved these fields directly into IDossierTemplate
we need to migrate the stored values in the annotations.
We also cleanup the storage by removing all the unused
IDossier fields in the annotation storage for DossierTemplates.