Prevent setting invalid reference prefix number.
I've decided to implement validation in the deserializer as it has all
the information necessary.

- It knows whether we are adding or editing
- It has all the data, context and parent can be accessed easily

Unfortunately an invariant cannot be used as it does not know whether we
are adding or editing. Also the add form does provide different data
during add than the rest api. In the form context is unavailable while
adding, while in the rest API the object to be validated is already
created and passed as as context.