Make sure to keep is_subdossier up to date for all types supporting it.
The event handler keeping is_subdossier up to date when a new object
is created or moved was only doing so for objects providing
the IDossierMarker interface. This is not really a problem because
DossierTemplates cannot be moved. Nevertheless, to avoid issues
in the future, it is safer to keep the index up to date for all
objects supporting it.