Don't prevent adding document with docprop issues.
No longer prevent adding documents with doc-property update issues.
There are currently problems with python-docx when it attempts to parse
bookmark references, which seem to be valid in word but will cause
an error when the document is parsed.

In any case the check we perform can easily be side-stepped by adding
an empty word and then replacing its contents when in checked out state.