Move validation to PloneParticipationHandler.add_participation.
Validating the participant_id and whether there is already
a participation for that participant_id on the given context
is now done in the PloneParticipationHandler. This means
that validation also happens when adding directly over the
handler. This forced us to port some tests to solr integration
tests, as contacts are queried over solr by the source used
for validation.