Raise when trying to delete a missing plone participation.
The PloneParticipationHandler should also raise MissingParticipation
when trying to delete a participation that does not exist
on the current context. The "quiet" argument was never used
(so always set to True), so that no exception would be raised
here. I don't see why we should not raise when trying to delete
an object that does not exist.