Only show relevant facets for participants and participation_roles.
In the listing endpoint, we not all facets from the participations
index are relevant for the participants and participation_roles
fields. Instead we show only the one facet for each role in
the participation_roles field (and not one per role and per user),
and one facet per participant in the participants endpoint.