Adapt test for REST API serialization of ReadOnlyErrors:
Nice we're now starting to not make write actions impossible during
read only mode by withdrawing the respective permissions from
users (in RO mode), the tests attempt to add a document will now
run into a 401 Unauthorized (instead of a ReadOnlyError).

We therefore adapt the test to hit a service specifially designed
for this.