Adapt test for ReadOnlyError error page:
Nice we're now starting to not make write actions impossible during
read only mode by withdrawing the respective permissions from
users (in RO mode), the test's attempt to add a document will now
run into a 401 Unauthorized (instead of a ReadOnlyError).

We therefore adapt the test to hit a view specifially designed
for this.