Expose bumblebee notifications url in the config endpoint
The bumblebee notifications url is necessary for the gever-ui to open a websocket connection to bumblebee in order to receive updates about newly created previews. With those updates, the gever-ui is able to reload the document's preview images as soon as the preview is available in bumblebee.

See https://4teamwork.atlassian.net/browse/NE-80