Remove limit for Subject facet
In https://4teamwork.atlassian.net/browse/NE-173, the user reported an issues, where not all facets are listed in the dossier listing.
This is becuase the default facet limit is 100 in solr 8.4: https://lucene.apache.org/solr/guide/8_4/faceting.html#field-value-faceting-parameters.
So, remove the limit of the Subject facets by including f.Subject.facet.limit=-1 in the solr query.