Store references to patched method a patch time.
When patching methods at runtime in write on read tracing we have to
get the reference to the patched methods at patch-time not a module load
time. As modified out of sync monkey-patches the same method
(Connection.register) we have to make sure that we don't unpatch the
patched method when removing write on read tracing patches.

Thus we now set the module-global reference to the original method in a
deferred way at patch time. Some extra paranoia is added in form of
additional assertions to make sure we don't set method references to
`None` by mistake. Also patching and unpatching is now protected by
an additional lock to prevent simultaneous requests in separate threads
from making a mess. To clarify entry points for clients the two
separate patching and unpatching methods have been inlined into one
method each.