Fix an encoding error on the local contacts tab.
We fix an encoding issue that arose due to inconsistent handling of
unicode vs bytestrings.

Items passed to the helper methods can be anything from a brain to a
solr-document and those do not gurantee unicode strings. Thus we
consider these methods to be at the application boundary and make sure
we work with unicode internally.

We also refactore the two nearly identical methods and merge them
into one method with two shallow wrapper methods setting correct
parameters.