Return retrieval mode ('copy' or 'version') when retrieving docs from workspace:
This is the mode that was finally picked (by the system) to retrieve a document
from a workspace back to GEVER. This is not necessarily the same as was requested
via the `as_new_version` flag.

This helps making better test assertions, and is included in the serialized
response so that eventually it could be used to display this to the user, in
order to let them know which mode was selected for their individual documents.