Use form action URL to re-render specific widget.
Instead of using browser location we use the form action to re-render
specific widgets. The browser location does not always expose the add
form name, for example when opened via POST from a tabbed view action.

The form action however always points to the correct form and is more
reliable.

This fixes a bug with only rendering allowed proposal templates for the
chosen committee on a proposal add form, when the form was opened with
the "add proposal" action on a document tab in the classic ui.