Fix deletion of favorites when object is removed.
When an object is removed, the corresponding favorites get deleted.
This did not work because the OGUID of the object being removed
cannot be determined anymore on the ObjectRemoved event, as the
object has already been removed from the IntId catalog. We now
register that handler on the WillBeRemoved event instead.