No longer globally omit custom_properties field.
This is necessary as `plone.restapi` relies on an ad-hoc form to
generate the schemas for the FTI. That ad-hoc form also applies any form
hints placed in the behavior schemas.

We push the form hints down into the add and edit forms respecitevly.
This is only a temporary workaround so we can move forward with custom
properties in API `@schema` endpoint without proper classic UI support.

Once we implement support in the classic UI we can get rid of this.