Properly translate 'label_name' in sharing tab:
Our sharing tab uses the msgid 'label_name' as a column header,
and translates it in the 'plone' domain. That domain has a German
translation in plone.app.locales, but for some reason the English
one is empty.

We therefore simply include it in our overridden translations
(for all three languages) for the plone domain.