Drop link label "docs" in favour of target title.
I know that the docs are german and thus we will mix german and english,
but i'm expecting the german docs to go away. Having the actual title
of the target in the links improves navigation and helps the reader to
see which links go to the same target and which do go to different
pages.