Move API changelog to top of index.
The changelog is a bit hidden at the moment. While that may be ok for a
stable API I'm of the opinion that with our API the changes should be a
bit more accessible/visible. Going through recent changes/additions to
the API may be the dominating use-case for visiting the docs.