Fix choice fields always skipping validation.
Fields are constructed by `ChoiceHandler`, a choice-field specific
`IFieldExportImportHandler` implementation. It does not seem to
construct the fields via their constructor and thus never sets
the `_init_field` instance variable to `False` to signal that
initialization is complete. This will cause the field to always skip
validation.

To work around this issue we set the attribute manually after the
schema class and its fields have been loaded or a field is added.