Don't warn about failed email notifications for recipients without an email address:
If a notification recipient's OGDS user doesn't have an email address, no attempt
should be made at dispatching an email (even if they have email notifications turned on).
As a consequence of this, the activity's actor will not see a warning any more that
one or more notifications couldn't be delivered.