Mail without a message is invalid.
The widget for the message field of mails always gets displayed in
the NO_DOWNLOAD_DISPLAY_MODE, which leads to an error when there is
no message. As a Mail without a message is anyway not valid, this
is a problem only in tests, for mails constructed with the builders.