Raise a specific exception for unmapped activity kinds:
If an activity kind is not mapped at all to a notification
settings group at all, something went wrong and we'll want to know.

A specific exception should therefore be raised for this. When using
the higher level add_activity() method on the PloneNotificationCenter
though, it will be caught though and won't prevent the attempted
operation - it will just be logged and reported to Sentry.