Allow any authenticated users to use the REST API.
This PR solves the problem that if a user in the new frontend,
i.e. via API, wants to access a foreign task in a foreign admin unit.

With the current implementation, the task responsible is allowed to
access the task on the foreign admin_unit, but not to use the API on
PloneSite root level, wich leads in Unauthorized Exceptions when using
the new frontend and clicking on a foreign task.

https://4teamwork.atlassian.net/browse/CA-1859