Fix DeactivatedCatalogIndexing context-manager.
DeactivatedCatalogIndexing was not working properly because collective.indexing
was still adding indexing operations to the indexing queue, and the patched
methods were really only used when the queue was processed. This led to issues
as whether an object would be indexed or not depended on when the queue was
processed instead of when the object was added to the queue.
To fix that issue we modify the Monkey patch to always skip indexing when
active and modify the context-manager to patch and unpatch the indexing
methods instead of adding an interface to the request as was done
previously.