Don't attempt to update remote timestamps during OGDS sync in
TTW bundle import. With the current order of operations, this
can't work because the admin unit for the local site hasn't been
created yet, but the admin unit would be required to authenticate
the request.

Another way to address this could be to not sync OGDS immediately
after LDAP setup, but wait until admin units have been created as
well. That would require dropping the strict existence check for
org units' user groups though.